refactor: mv useCurrency logic to lib/hooks
